Proposed fix for Unown
When an Unown is found in the wild, the game hangs until it generates DVs that yield an Unown letter that the player has unlocked.  If the player has not completed an Alph puzzle to unlock Unown (as is frequently the case in randomizers), the game will loop infinitely.  As a result of this fix, any Unown letter can be generated in the wild at any time, regardless of whether the player has completed an Alph puzzle.